BUILD-4602 Move releasability_check tests within their dedicated test file (#145)
That way they are where they belong
